requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
torch>=1.10.0
torchvision>=0.11.0
Pillow>=9.0.0
opencv-python>=4.5.0
scikit-image>=0.18.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0
seaborn>=0.11.0

kivy>=2.0.0